prima partita 9a0, pre seconda partita

+++ src/vars.h
@@ -241,7 +241,7 @@ extr bool XP_SX;
 extr bool XP_DX;
 extr int portx;
 
-extr int stincr;
+extr float stincr;
 
 extr int fpos;
